#2908 Move test and improve it
#2908 Provide Doc Action voiding for contract period
